Update Template Literals
Moved Tagged Template Literals into own deck
Might not cover them - the use cases for them are not that obvious, and end up not being talked about later on in the training.

It is how things like styled components, htm, etc work under the hood.

I did wonder how

```
let Button = styled.a` .....`
```

worked before - but not sure how much value it has spending time on this for this training.