Lift RobotState
I know I'll want to use some of the RobotData on the dashboard

Moved the robot list / reducer to the top level container, and pass down the robot and the dispatcher to update them to the child.

Building the robot so it has the extra props on it is going to be consistent across the app - so we leave the building here for now.

The filtering logic is specific to the RobotSearch - so we leave that in the RobotSearch