Fix robot update bug
We were updating the jobTitleId on the robot - but not the text along with it.

If we changed the jobTitle - the list would show the old jobTitle, but other values that were updated were reflected.

Initial pass at the fix - when saving a robot in the reducer, generate the IDs.

Although this  points out the value of the concept of "don't store derived state" - as every time we update a value that we are calculating something from -we need to think "where else do I need to update it?"

Will explore a better fix to this soon