SelectList - let consumer specify extra options
The SelectList was also taking control over rendering the 'Select One', or 'Any' type of option.

So on the search screen, and in the robot form were displaying

```jsx
 <option value={-1}>Select One</option>
````

We want the consumer to control how to display this, and if we want to display it or not.

We could in theory push an item into the options passed in and modify the array, but that feels clunky