Setup Hamburger Menu
The Bulma navbar is responsive by default and will collapse down to a hamburger menu on smaller screen.

The Bulma documentation gave this example to setup the toggle using vanilla js

```js
document.addEventListener('DOMContentLoaded', () => {

  // Get all "navbar-burger" elements
  const $navbarBurgers = Array.prototype.slice.call(document.querySelectorAll('.navbar-burger'), 0);

  // Check if there are any navbar burgers
  if ($navbarBurgers.length > 0) {

    // Add a click event on each of them
    $navbarBurgers.forEach( el => {
      el.addEventListener('click', () => {

        // Get the target from the "data-target" attribute
        const target = el.dataset.target;
        const $target = document.getElementById(target);

        // Toggle the "is-active" class on both the "navbar-burger" and the "navbar-menu"
        el.classList.toggle('is-active');
        $target.classList.toggle('is-active');

      });
    });
  }

});
```

We could copy/paste this into the index.html and it would work - but we are using React, so lets leverage react.

The SiteHeader was not using any state at all before - but now we need to keep track of if the menu is expanded or not.

We will use that to drive if the `is-active` class is applied

We use the `onClick` on the burger menu icon to toggle if it's expanded or not - and use the `useState` to keep track of the state